requirements.txt
fastapi
uvicorn
tinydb
streamlit
openai
requests
python-dotenv
google-genai